chore: reorganize demo into dedicated folder
