+++ src/demo/palettes.js
